+++ hakaiApi/R/client.R
@@ -72,18 +72,19 @@ Client <- R6::R6Class("Client",  # nolint
     #'@description
     #' Send a GET request to the API
     #' @param endpoint_url The full API url to fetch data from
+    #' @param col_types a readr type specification  
     #' @return A dataframe of the requested data
     #' @examples
     #' try(client$get("https://hecate.hakai.org/api/aco/views/projects"))
-    get = function(endpoint_url) {
+    get = function(endpoint_url, col_types = NULL) {
       token <- paste(private$credentials$token_type,
                      private$credentials$access_token)
-      r <- httr2::request(endpoint_url) |>
-        httr2::req_headers("Authorization" = token) |>
+      r <- base_request(endpoint_url, token) |> 
         httr2::req_perform()
+      data <- httr2::resp_body_json(r)
       data <- private$json2tbl(httr2::resp_body_json(r))
       data <- tibble::as_tibble(data)
-      data <- readr::type_convert(data)
+      data <- readr::type_convert(data, col_types = col_types)
       return(data)
     },
 
@@ -95,8 +96,7 @@ Client <- R6::R6Class("Client",  # nolint
     post = function(endpoint_url, rec_data) {
       token <- paste(private$credentials$token_type,
                      private$credentials$access_token)
-      resp <- httr2::request(endpoint_url) |>
-        httr2::req_headers("Authorization" = token) |>
+      resp <- base_request(endpoint_url, token) |>
         httr2::req_method("POST") |>
         httr2::req_body_json(rec_data) |>
         httr2::req_perform()
@@ -112,8 +112,7 @@ Client <- R6::R6Class("Client",  # nolint
     put = function(endpoint_url, rec_data) {
       token <- paste(private$credentials$token_type,
                      private$credentials$access_token)
-      resp <- httr2::request(endpoint_url) |>
-        httr2::req_headers("Authorization" = token) |>
+      resp <- base_request(endpoint_url, token) |>
         httr2::req_body_json(data = rec_data, auto_unbox = TRUE) |>
         httr2::req_method("PUT") |>
         httr2::req_perform()
@@ -129,8 +128,7 @@ Client <- R6::R6Class("Client",  # nolint
     patch = function(endpoint_url, rec_data) {
       token <- paste(private$credentials$token_type,
                      private$credentials$access_token)
-      resp <- httr2::request(endpoint_url) |>
-        httr2::req_headers("Authorization" = token) |>
+      resp <- base_request(endpoint_url, token) |>
         httr2::req_body_json(data = rec_data, auto_unbox = TRUE) |>
         httr2::req_method("PATCH") |>
         httr2::req_perform()
@@ -155,12 +153,7 @@ Client <- R6::R6Class("Client",  # nolint
     credentials_file = NULL,
     credentials = NULL,
     json2tbl = function(data) {
-      data <- lapply(data, function(data) {
-        data[sapply(data, is.null)] <- NA  # nolint
-        unlist(data)
-      })
-      data <- bind_rows(data)
-      return(data)
+      json2tbl_impl(data)
     },
     querystring2df = function(querystring) {
       tryCatch({