+++ hakaiApi/R/utils.R
@@ -1,7 +1,9 @@
 base_request <- function(endpoint_url, token) {
+  user_agent <- Sys.getenv("HAKAI_API_USER_AGENT", "hakai-api-client-r")
+  
   httr2::request(endpoint_url) |>
     httr2::req_headers("Authorization" = token) |>
-    httr2::req_user_agent("hakai-api-client-r") 
+    httr2::req_user_agent(user_agent) 
 }